+++ ui/components/radio-group/RELEASENOTES.md
@@ -10,6 +10,7 @@
 
 - Added styling changes to support radio group with help text icon.
 - Added examples: Required with help text icon and Right to left required with help text icon to documentation
+- Added required with help text icon and tooltip example
 
 ### Fixed